Make ConfigEntry.key and ConfigEntry.value private
They are now stored in __key and __value respectively.
Instead the key and value are stored by the get_key()
and get_value() methods.

This is done in order to process these entries before returning them
a use case for this is data="path", notice path is in quotes, for most
cases we need the path to not be in quotes so get_value() for data
entries returns a value with the quotes stripped unless raw=True is
passed along the method arguments.

Also refactored ConfigEntry.__init__()