Refactor core.get_mod_entry() to take a OmwMod object
Instead of a string, this allows to simplify the function a great deal.
As a side effect lib.config is no longer needed for lib.core
so i can safely remove it and start using core functions inside config.

Also updated the uninstall_mod function to adapt to the change.