Update lib.config to use core.get_full_path()
Instead of defining its own function.
This used to be a problem because previously core would import config
and config would.
This problem should be prevented in the future. As a rule of thumb core
should never import any of the projects modules, if it has to then
something went wrong :/