Refactor ConfigFile to remove useless methods.
Lots of methods that emulated a list are removed, there is no need to
treat that class like a list and doing so may allow bugs to creep in.
The methods that remained are improved a bit.

__contains__ raises an error if the argument is not a ConfigEntry
instance.
__iter__ uses *args **kwargs :)

_get_lines() was removed aswell, it was really pointless.
__str__ has been renamed to tostring, i can imagine where that method
may cause problems

Also made use of the with statement when opening files.