Implement ConfigFile.add_entry() method.
Unlike append this method will insert the entry based on its key.
If matching keys already exist in the config then the entry will be
appended behind them, otherwise it will be inserted in the bottom of the
file.
This method should be called as much as possible instead of append
(infact append should probably be removed?)

The new method is now used in the install_mod function and the enable method for plugins
Also removed the now useless get_latest_key_index  and insert_data_entry
functions from core