Refactor ConfigEntry class.
This created more problems than it solved but that class was so bad it
had to be done.

__init__ looks cleaner now
_parse_line now handles empty lines aswell
lots of useless functions are removed with the exceptions of __str__ and
__eq__ which have been changed a bit
__str__ handles blank lines and also raises an error if it can't
determine that type of the entry.
__eq__ checks the key and value of both entries instead of using
str(self) == str(other)

ConfigFile has also been modified to support the new blank line entries.
In this process i discovered that it didn't previously handle comment
entries in any special manner. Now ConfigFile._entries store all entries
(for writing back to disk later) while get_entries returns only setting
entries, but with the all flag can be made to return all entries.