Add EsmTES3Record class and refactor how Esm classes.
Added EsmTES3Record that parses and modifies the first record in an esm
file, this record contains info about the author, description and
dependencies of that file.

The refactor was done because of an initial design flaw in the EsmRecord
class, it stored its data and parsed it into subrecords and stored that
too. The design flaw was that both data and subrecords are equivalent so
when packing and unpacking it was hard to figure out if the user changed
the raw data or a subrecord.

Now subrecords are no longer parsed automatically or stored.

- get_subrecords() will generate a new list of subrecords on the fly by
  calling get_data()

- get_data() will return the records data, if self._changed is True it
  will call pack_data()

- pack_data() a new method, in EsmRecord it returns self.__data, since
  that class is not expected to modify itself, however subclasses that
  do modify their data are expected to override this method to return
  their new data properly.

- unpack(), all previous methods that unpacked data are now renamed to
  unpack(), classes affected are Esm and EsmRecord.

- pack() same as unpack(), this function is generic in EsmRecord so no
  need to override it in a subclass, it will call get_data() so it
  expects up to date data.

- EsmLEVRecord.pack_sub() has been renamed to pack_subrecord() and
  moved to EsmRecord so its available to all subclasses, it has also
  been modified to accept a tuple of strings instead of a single string
  so you can pack multiple properties at once by specifying a proper
  format.