Add supports for blank lines and comments in openmw.cfg
Config entries are now split into 2 classes, the old ConfigEntry class
which was changed to drop support for comments and blank lines.
and ConfigRawEntry which basically holds the raw line straight from the
config and the config object in case its ever needed.

Making ConfigEntry a subclass of ConfigRawEntry is work for another
time.

Also ConfigFile was being treated as list (its not) so i fixed that.
no more __iter__ or __contains__, if you want to check the contents of
the config file use get_config().