Refactor ConfigFile, OmwMod and OmwPlugin
ConfigEntry are no more, they are only used to store config entries we
dont care about.

OmwMod and OmwPlugin will now directly represent a data, content entry
This is to simplify modifications to ConfigFile and remove the need to
keep the object and entry in sync with each other.

Due to this handling of plugins is now improved, functions that can list
enabled, disabled and orphaned plugins exist, plugin load order can be
rearranged be simply moving them inside ConfigFile. The same thing will
be done for OmwMod in the future.